+++ CHANGELOG.md
@@ -10,6 +10,7 @@ For more information about changelogs, check
 
 * [FEATURE] Add `icon` helper
 * [FEATURE] Add `font_awesome_css` helper
+* [FEATURE] Add `dropdown` helper
 
 ## 1.0.1 - 2014-09-14